Update checkout_complete.py
Changed the algorithm. Instead of getting a randomized item, I direct it to a specific item that always have a stock, so that add to cart will be triggered always. try/except will be useless because of it but I like to keep it for reference. Also added checkout with Cash on Delivery transaction.